работа над ошибками

+++ classes/link.php
@@ -2,43 +2,42 @@
 
 class Link {
 
-    public $response = [];
-    private $link;
+    public $response = [
+        "status" => true,
+        "message" => "",
+    ];
     private $shortLinkMaxLen = 8;
     private $shortLinkMaxCount;
     private $alphabet = [];
     private $alphabetCount;
     private $table_names = [];
+    private $redirectLink;
 
-    public function __construct($table_names, $db_conn)
+    public function __construct($table_names, $db_conn, $redirectLink)
     {
         $this->table_names = $table_names;
         $this->db_conn = $db_conn;
-        $this->response["status"] = false;
-        $this->response["message"] = "";
         $this->alphabet = array_merge(range('A', 'Z'), range('a', 'z'), range(0, 9));
         $this->alphabetCount = count($this->alphabet);
         $this->shortLinkMaxCount = $this->gmp_fact($this->alphabetCount) / ( $this->gmp_fact($this->alphabetCount - $this->shortLinkMaxLen) - $this->gmp_fact($this->shortLinkMaxLen) ); // Максимально возможное число коротких ссылок (факториал по общему количеству элементов делим на факториал разницы общего числа элементов с максимальным числом, умноженным на факториал максимального числа ( n! / (n-k)! * k! ))
+        $this->redirectLink = $redirectLink;
     }
 
     // Создание короткой ссылки
     public function createShortLink($link) {
-        $this->encodeLink($link);
-        if ($this->response["status"] == "true") {
-            $shortLink = $this->generateShortLink();
+        $link = $this->encodeLink($link);
+        if ($link !== false) {
             $i = 0;
-            $this->response["status"] = "true"; // Устанавливаем статус в true, чтобы можно было отловить момент неудачной операции с бд при проверке
-            // Обходим циклом до тех пор, пока ссылка не будет валидной, пока установлен положительный статус (получается подключаться к бд), и пока не переберется максимально возможное число ссылок для заданной длины
-            while (!$this->checkShortLink($shortLink) && $this->response["status"] == "true" && $i < ($this->shortLinkMaxCount - 1)) {
+            $shortLink = $this->generateShortLink();
+            while (!$this->checkShortLink($shortLink) && $this->response["status"] !== false && $i < ($this->shortLinkMaxCount - 1)) {
                 $shortLink = $this->generateShortLink();
                 $i++;
             }
-            if ($i == $this->shortLinkMaxCount - 1)
-                $this->setResponce("false", "Достигнуто максимально доступное количество ссылок, обратитесь к администратору");
-            else {
-                $this->saveShortLink($this->link, $shortLink);
-                if ($this->response["status"] == "true")
-                    $this->setResponce("true", "Созданная короткая ссылка: " . $this->getHTMLLink($shortLink));
+            if ($this->response["status"] !== false && $i == $this->shortLinkMaxCount - 1)
+                $this->setResponce(false, "Достигнуто максимально доступное количество ссылок, обратитесь к администратору");
+            else if ($this->response["status"] !== false) {
+                if ($this->saveShortLink($link, $shortLink))
+                    $this->setResponce(true, "Созданная короткая ссылка: " . $this->getLinkHtml($shortLink));
             }
         }
     }
@@ -59,20 +58,21 @@ class Link {
         if ($result) {
             if ($result->num_rows == 0)
                 return true;
-            else
-                return false;
         } else
-            $this->setResponce("false", "Произошла ошибка базы данных при обработке запроса проверки наличия: " . $this->db_conn->error);
+            $this->setResponce(false, "Произошла ошибка базы данных при обработке запроса проверки наличия: " . $this->db_conn->error);
+        return false;
     }
 
     // Сохранение полученной короткой ссылки
     private function saveShortLink($link, $shortLink) {
         $query = "INSERT INTO {$this->table_names["links"]} (link, short_link) VALUES('{$link}', '{$shortLink}')";
         $result = $this->db_conn->query($query);
-        if ($result)
-            $this->setResponce("true");
-        else
-            $this->setResponce("false", "Произошла ошибка базы данных при обработке запроса сохранения: " . $this->db_conn->error);
+        if ($result) {
+            $this->setResponce(true);
+            return true;
+        } else
+            $this->setResponce(false, "Произошла ошибка базы данных при обработке запроса сохранения: " . $this->db_conn->error);
+        return false;
     }
 
     // Получение всех записей
@@ -90,59 +90,71 @@ class Link {
                 }
                 return $links;
             } else
-                $this->setResponce("false", "Список пуст.");
+                $this->setResponce(false, "Список пуст.");
         } else
-            $this->setResponce("false", "Произошла ошибка базы данных при обработке запроса получения: " . $this->db_conn->error);
+            $this->setResponce(false, "Произошла ошибка базы данных при обработке запроса получения: " . $this->db_conn->error);
         return false;
     }
 
     // Получение полной ссылки
     public function getFullLink($short_link) {
-        if ($short_link) {
+        $short_link = $this->validate($short_link);
+        if ($short_link !== false) {
             $query = "SELECT link FROM {$this->table_names["links"]} WHERE short_link = '{$short_link}'";
             $result = $this->db_conn->query($query);
             if ($result) {
                 if ($result->num_rows > 0) {
                     $result = $result->fetch_assoc();
                     $result = $result["link"];
-                    $this->setResponce("true", "Ваша полная ссылка: " . $result);
+                    $this->setResponce(true, "Ваша полная ссылка: " . $result);
                 } else
-                    $this->setResponce("false", "Введенной ссылки не существует");
+                    $this->setResponce(false, "Введенной ссылки не существует");
             } else
-                $this->setResponce("false", "Произошла ошибка базы данных при обработке запроса получения: " . $this->db_conn->error);
-        } else
-            $this->setResponce("false", "Короткая ссылка пуста");
+                $this->setResponce(false, "Произошла ошибка базы данных при обработке запроса получения: " . $this->db_conn->error);
+        }
     }
 
-    // Смена ссылки по её короткому коду
+    // Смена ссылки по её короткому коду (Почему-то метод affected_rows не определялся, сделал по простому)
     public function changeLink($link, $shortLink) {
-        $query = "UPDATE {$this->table_names['links']} SET link = '{$link}' WHERE short_link = '{$shortLink}'";
-        $result = $this->db_conn->query($query);
-        if ($result) {
-            // Почему-то affected_rows не определялся
-//            $changedRowsCount = $result->num_rows();
-//            if ($changedRowsCount > 0)
-                $this->setResponce("true", "Запись успешна обновлена, новое значение - " . $link . ", по короткой ссылке - " . $shortLink);
-//            else
-//                $this->setResponce("false", "Не найдена информация по данному запросу");
-        } else
-            $this->setResponce("false", "Произошла ошибка базы данных при обработке запроса изменения: " . $this->db_conn->error);
+        $link = $this->validate($link);
+        if ($link !== false)
+            $shortLink = $this->validate($shortLink);
+                if ($shortLink !== false) {
+                    $query = "UPDATE {$this->table_names['links']} SET link = '{$link}' WHERE short_link = '{$shortLink}'";
+                    $result = $this->db_conn->query($query);
+                    if ($result)
+                        $this->setResponce(true, "Запись успешна обновлена, новое значение - " . $link . ", по короткой ссылке - " . $shortLink);
+                    else
+                        $this->setResponce(false, "Произошла ошибка базы данных при обработке запроса изменения: " . $this->db_conn->error);
+                }
     }
 
-    // Удаление ссылки (Можно было бы использовать affected_rows, но он не работает
+    // Удаление ссылки (Почему-то метод affected_rows не определялся, сделал по простому)
     public function removeLink($id) {
-        $query = "DELETE FROM {$this->table_names["links"]} WHERE id = {$id}";
-        $result = $this->db_conn->query($query);
-        if ($result)
-            $this->setResponce("true", "Запись успешно удалена");
-        else
-            $this->setResponce("false", "Произошла ошибка базы данных при обработке запроса удаления: " . $this->db_conn->error);
+        $id = htmlspecialchars(strip_tags($id));
+        if (strlen($id) > 0) {
+            $id = intval($id);
+            if ($id >= 0) {
+                $query = "DELETE FROM {$this->table_names["links"]} WHERE id = {$id}";
+                $result = $this->db_conn->query($query);
+                if ($result)
+                    $this->setResponce(true, "Запись успешно удалена");
+                else
+                    $this->setResponce(false, "Произошла ошибка базы данных при обработке запроса удаления: " . $this->db_conn->error);
+            } else
+                $this->setResponce(false, "Число должно быть больше или равно 0");
+        } else
+            $this->setResponce(false, "Введено некорректное значение");
     }
 
-    // Для удобства возврата ответа
     private function setResponce($status, $message = "") {
         $this->response["status"] = $status;
         $this->response["message"] = $message;
+    }
+
+    public function returnResponse() {
+        $statusClass = $this->response["status"] ? "text-success" : "text-error";
+        return "<p class='" . $statusClass . "'>" . $this->response["message"] . "</p>";
     }
 
     // Вычисление факториала (на опенсервере не было модуля, описал отдельно)
@@ -153,26 +165,48 @@ class Link {
         return $factorial;
     }
 
-    // Валидатор
     public function validate($link) {
         $link = htmlspecialchars(strip_tags($link));
-        return $link;
+        if (!$link)
+            $this->setResponce(false, "Ссылка вернулась пустой");
+        else if (strlen($link) > 100)
+            $this->setResponce(false, "Ссылка не должна превышать 100 символов");
+        else {
+            $this->setResponce(true);
+            return $link;
+        }
+        return false;
     }
 
-    // Валидатор и преобразование
     private function encodeLink($link) {
         $link = $this->validate($link);
-        if ($link) {
-            $this->link = urlencode($link);
-            $this->setResponce("true");
-        } else
-            $this->setResponce("false", "Не верно задана ссылка");
+        if ($link !== false) {
+            $this->setResponce(true);
+            return urlencode($link);
+        }
+        return false;
     }
 
-    // Получение готовой ссылки для перехода с версткой
-    public function getHTMLLink($shortLink) {
-        $location = $_SERVER["DOCUMENT_ROOT"] . "/?l=" . $shortLink;
-        return "<a href='$location'>{$location}</a>";
+    public function getLinkHtml($shortLink) {
+        $link = $this->redirectLink . "?l=" . $shortLink;
+        return "<a href='$link'>{$link}</a>";
+    }
+
+    // Получение ссылки по короткой и редирект на неё
+    public function getLinkAndRedirect($shortLink) {
+        $query = "SELECT link FROM {$this->$table_names['links']} WHERE short_link = '{$shortLink}'";
+        $result = $this->db_conn->query($query);
+        if ($result) {
+            $result = $result->fetch_assoc();
+            $result = $result["link"];
+            if ($result) {
+                $this->setResponce(true);
+                header("Location: " . $result);
+            } else
+                $this->setResponce(false, "Произошла ошибка, вернулась пустая ссылка");
+        } else {
+            $this->setResponce(false, "Произошла ошибка базы данных при обработке запроса: " . $this->db_conn->error);
+        }
     }
 
 }
